% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTCF_hg19}
\alias{CTCF_hg19}
\title{Predicted CTCF-motif binding sites}
\format{
a GRanges object
}
\usage{
data("CTCF_hg19")
}
\description{
Genomic coordinates of predicted CTCF binding sites with motif MA0139.1 (Jaspar), in BED format, with strand orientation (directionality of binding). Human (hg19, hg38) and mouse (mm9, mm10) genomes. The binding sites were detected using the FIMO tool of the MEME suite using default settings. Extra columns include motif name (MA0139.1), score, p-value, q-value, and the motif sequence.
}
\details{
Experimental, to be submitted as an AnnotatiohHub Bioconductor package.
}
\references{
Dozmorov MG (2021). \emph{CTCF}. https://github.com/mdozmorov/CTCF/CTCF - R
package version 0.99.0, <URL: https://github.com/mdozmorov/CTCF>.
}
\keyword{datasets}
