% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deny}
\alias{deny}
\title{ENCODE Unified GRCh38 Exclusion List (previously "blacklist")}
\format{
a GRanges object
}
\usage{
data("deny")
}
\description{
The blacklist is a unification of three separate blacklists followed by extensive manual curation generated by Anshul Kundaje, Stanford https://www.encodeproject.org/files/ENCFF356LFX/
The three blacklists that were unified are as follows
\enumerate{
\item The autogenerated blacklist from https://github.com/Boyle-Lab/Blacklist/blob/master/lists/hg38-blacklist.v2.bed.gz .
The corresponding citation is
Amemiya, H.M., Kundaje, A. & Boyle, A.P. The ENCODE Blacklist: Identification of Problematic Regions of the Genome. Sci Rep 9, 9354 (2019). https://doi.org/10.1038/s41598-019-45839-z
\item The autogenerated peakPass blacklist from https://github.com/ewimberley/peakPass/blob/master/blacklists/hg38/peakPass60Perc_sorted.bed
The corresponding citation is
Charles E. Wimberley and Steffen Heber.Journal of Computational Biology.Feb 2020.259-268.http://doi.org/10.1089/cmb.2019.0295
\item The blacklist capturing peaks in the nuclear genome that originate from reads from mitochondrial genome taken from https://github.com/caleblareau/mitoblacklist/blob/master/peaks/hg38_peaks.narrowPeak . This blacklist was generated by Caleb Lareau.
}
}
\keyword{datasets}
