% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHSA549Hg38}
\alias{DHSA549Hg38}
\title{nullranges DNase hypersensitivity peaks in A549 cell example data}
\format{
a GRanges object with metadata score, signal value, p/q value and peak.
}
\usage{
data("DHSA549Hg38")
}
\description{
An example dataset containing narrowPeak file from ENCODE. Retrieve record with \code{object[["AH22505"]]} on Annotation Hub. Analysis script is in 'inst/script/DHSA549Hg38.R'.
}
\keyword{datasets}
