% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enhPromContactFreqHg19_v2}
\alias{enhPromContactFreqHg19_v2}
\title{nullranges enhancer-promoter contact frequency example data}
\format{
a GInteractions object with metadata contact frequencies and E-P distance.
}
\usage{
data("enhPromContactFreqHg19_v2")
}
\description{
An example dataset containing the interaction frequencies between all putative enhancer-promoter (E-P) pairs within 2Mb. Promoter regions from the monoMacroDEGHg19 dataset and enhancer regions from the H3K27acPeakCountsHg19 dataset were paired within 2Mb regions around gene promoters. E-P pairs were then assigned to the appropriate Hi-C bin at 10kb resolution, taking the center point of each enhancer and the TSS of each promoter. Interaction frequencies between each E-P pair were extracted from bioreps of monocyte (NON_1_hic, NON_2_hic) and macrophage (PMA_1_hic, PMA_2_hic) Hi-C files. Total contact frequency and E-P distance were also calculated. Returned is a GInteractions object with all putative E-P pairs along with their contact frequencies. Analysis script is in 'inst/script/enhPromContactFreqHg19_v2.R'.
}
\details{
The major difference between enhPromContactFreqHg19 and enhPromContactFreqHg19_v2 is the way "2Mb around the promoter" is defined. In enhPromContactFreqHg19, 2Mb is expanded around the center of each promoter - where the promoter is defined as 2kb upstream and 200bp downstream from the TSS. In enhPromContactFreqHg19_v2, 2Mb is expanded around the TSS of each gene, rather than some point in the middle of the promoter.
}
\references{
Phanstiel, Douglas H., Kevin Van Bortle, Damek Spacek, Gaelen T. Hess, Muhammad Saad Shamim, Ido Machol, Michael I. Love, Erez Lieberman Aiden, Michael C. Bassik, and Michael P. Snyder. 2017. “Static and Dynamic DNA Loops Form AP-1-Bound Activation Hubs during Macrophage Development.” Molecular Cell 67 (6): 1037–48.e6.
}
\keyword{datasets}
