% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exclude}
\alias{exclude}
\title{ENCODE Unified GRCh38 Exclusion List}
\format{
a GRanges object
}
\usage{
data("exclude")
}
\description{
The Exclusion List is a unification of three separate lists
followed by extensive manual curation generated by Anshul Kundaje,
Stanford https://www.encodeproject.org/files/ENCFF356LFX/
}
\details{
Analysis script is in 'inst/script/exclude.R'.
}
\keyword{datasets}
