% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H3K27acPeakCountsHg38}
\alias{H3K27acPeakCountsHg38}
\title{nullranges H3K27ac monocyte/macrophage peak count example data}
\format{
a GRanges object with metadata peak counts
}
\usage{
data("H3K27acPeakCountsHg38")
}
\description{
An example dataset containing GRanges of merged H3K27ac ChIP-seq peaks in THP-1 cells either untreated (NON, monocytes) or treated with PMA (PMA, macrophages). Metadata columns contain peak counts from each condition and biological replicate (NON_1, NON_2, PMA_1, and PMA_2)  obtained from \code{bedtools multicov}. Additional metadata columns contain summary statistics calculated in "inst/script/H3K27acPeakCountsHg38.R". ChIP-seq data was aligned to hg38.
}
\references{
Phanstiel, Douglas H., Kevin Van Bortle, Damek Spacek, Gaelen T. Hess, Muhammad Saad Shamim, Ido Machol, Michael I. Love, Erez Lieberman Aiden, Michael C. Bassik, and Michael P. Snyder. 2017. “Static and Dynamic DNA Loops Form AP-1-Bound Activation Hubs during Macrophage Development.” Molecular Cell 67 (6): 1037–48.e6.
}
\keyword{datasets}
