% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monoMacroDEGHg19}
\alias{monoMacroDEGHg19}
\title{nullranges macrophage/monocyte differential gene example data}
\format{
a GRanges object with metadata peak counts
}
\usage{
data("monoMacroDEGHg19")
}
\description{
An example dataset containing the results of a DESeq2 analysis of THP-1 cells either untreated (NON, monocytes) or treated with PMA (PMA, macrophages). The dataset is a GRanges object containing the locations of genes and metadata columns contain the DESeq2 results. Analysis of raw data in "inst/script/monoMacroDEGHg19.R". RNA-seq was quantified with salmon-1.4.0 to hg19.
}
\references{
Phanstiel, Douglas H., Kevin Van Bortle, Damek Spacek, Gaelen T. Hess, Muhammad Saad Shamim, Ido Machol, Michael I. Love, Erez Lieberman Aiden, Michael C. Bassik, and Michael P. Snyder. 2017. “Static and Dynamic DNA Loops Form AP-1-Bound Activation Hubs during Macrophage Development.” Molecular Cell 67 (6): 1037–48.e6.
}
\keyword{datasets}
